Hachi.Compiler.CodeGen.CPS

Description

This module contains definitions related to continuations.

Types

contEntryTy :: Type

contEntryTy is the Type of continuation entry functions.

contTyDef :: Type

contTyDef is the definition of the continuation closure type.

Code generation

compileDynamicCont :: MonadCodeGen m => String -> Set (Text, Bool) -> Text -> (Operand -> Operand -> IRBuilderT m ()) -> IRBuilderT m Continuation

compileDynamicCont name freeVars parameter builder generates code which dynamically allocates a continuation closure.

Identity continuation

cpsReturnOp :: Operand

cpsReturnOp is a reference to the global identity continuation.

cpsReturnCont :: Continuation

cpsReturnCont is a reference to the global identity continuation.

compileCpsReturn :: MonadCodeGen m => m Continuation

compileCpsReturn is a computation which generates code for the identity continuation, which just returns its argument and makes no further function calls. This is used to terminate a CPS style sequence of function calls.

Calling continuations

callCont :: (ContinuationRep a, MonadCodeGen m, MonadIRBuilder m) => Continuation -> a -> m (ClosurePtr 'DynamicPtr)

callCont continuationPtr argument generates code which calls the continuation pointed at by continuationPtr with the argument.