Hachi.Compiler.CodeGen.Closure

Closure representation

data ClosurePtr (k :: PtrKind) where

Represents a pointer to a closure.

Constructors

MkClosurePtr :: Operand -> ClosurePtr 'DynamicPtr

MkStaticClosurePtr :: Constant -> ClosurePtr 'StaticPtr

closureTyDef :: Type

closureTyDef is a Type definition for closures. Note the layout is as follows:

A pointer to the code for the closure.

A pointer to the print code for the closure.

Closure flags.

Zero or more free variables. (LLVM will accept more elements than the array's indicated size) https://llvm.org/docs/LangRef.html#array-type

closureSize :: Num a => a

The minimum size of a closure in words.

closureTy :: Type

closureTy is a Type for closures.

closureTyPtr :: Type

closureTyPtr is a Type representing a pointer to a closure.

Creating closures

mkEntryTy :: Int -> Type

mkEntryTy arity generates a function signature for a closure with arity-many parameters in addition to the parameter that references the closure itself. In other words, the resulting signature has 1+arity many parameters.

clsEntryTy :: Type

clsEntryTy is the Type of closure entry functions.

clsEntryParams :: Text -> [(Type, ParameterName)]

clsEntryParams varName constructs a list of parameters for a closure's entry function, where varName is the name of the variable being bound.

varEntryTy :: Type

printFnTy :: Type

printFnTy is the Type of closure print functions. They are similar to closure entry functions in that they take a pointer to the current closure as argument, but return nothing.

mkClosureName :: String -> Name

mkClosureName name returns the name of a closure for name.

compileClosure :: MonadModuleBuilder m => Bool -> String -> Constant -> Constant -> [Constant] -> m (ClosurePtr 'StaticPtr)

compileClosure isPoly name entryPtr printPtr generates a global variable representing a closure for name.

ptrSize :: Operand

The size of a pointer as an Operand. This should correspond to the value given by bits.

bits :: Word32

When we dynamically allocate closures, we need to know how much space to allocate for the pointers; for this purpose we are currently assuming that we are running on a 64-bit system.

class HasFreeVars ptr where

A class of types which represent runtime structures that contain free variables.

Methods

loadFreeVar :: (MonadModuleBuilder m, MonadIRBuilder m) => ptr -> Type -> Integer -> m Operand

loadFreeVar ptr type index loads the free variable at index of the closure-like structure pointed to by ptr. If required, the resulting Operand is cast to type.

allocateClosure :: (MonadCodeGen m, MonadIRBuilder m) => Bool -> Constant -> Constant -> [Operand] -> m (ClosurePtr 'DynamicPtr)

allocateClosure isDelay codePtr printPtr freeVars allocates a closure with enough space to store all of freeVars in addition to the code pointers. The code pointers are free variables are stored in the closure and the pointer to the closure is returned.

withFreeVars :: (HasFreeVars ptr, MonadCodeGen m, MonadIRBuilder m) => ptr -> m a -> m a

withFreeVars ptr computation runs computation after retrieving all free variables from the closure-like structure pointed to by ptr.

compileDynamicClosure :: MonadCodeGen m => Bool -> String -> Set (Text, Bool) -> Text -> (Operand -> Operand -> Continuation -> IRBuilderT m ()) -> IRBuilderT m (ClosurePtr 'DynamicPtr)

compileDynamicClosure isDelay name freeVars var codeFun generates code which dynamically creates a new closure when executed. The static parts of the closure use name. The set given by freeVars signals which variables should be stored in the closure's free variables array. The name given by var is the variable that is bound by this closure. codeFun is a computation which generates the code for the body of the closure. The function returns a ClosurePtr representing the new closure.

compileFunPrint :: MonadCodeGen m => m Constant

compileFunPrint is a computation which generates the shared pretty-printing code for all function closures and returns a reference to it.

Loading data from closures

data ClosureComponent

Enumerates different components of a closure.

Constructors

ClosureCode

ClosurePrint

ClosureFlags

ClosureFreeVar Integer

loadFromClosure :: (MonadModuleBuilder m, MonadIRBuilder m) => ClosureComponent -> Type -> ClosurePtr k -> m Operand

loadFromClosure component type ptr loads the component described by component from the closure represented by ptr and casts its type to type.

callClosure :: (MonadFail m, MonadModuleBuilder m, MonadIRBuilder m) => ClosureComponent -> ClosurePtr k -> [Operand] -> m (ClosurePtr 'DynamicPtr)

callClosure component ptr args loads the component described by component from the closure represented by ptr, assumes it is a function, and then calls it with args as arguments.

enterClosure :: (MonadFail m, MonadModuleBuilder m, MonadIRBuilder m) => ClosurePtr k -> Continuation -> Maybe Operand -> m (ClosurePtr 'DynamicPtr)

printClosure :: (MonadFail m, MonadModuleBuilder m, MonadIRBuilder m) => ClosurePtr k -> m ()

printClosure ptr invokes the pretty-printing function of the closure represented by ptr.

compileApply :: (MonadFail m, MonadModuleBuilder m, MonadIRBuilder m) => Continuation -> ClosurePtr f -> ClosurePtr x -> m (ClosurePtr 'DynamicPtr)

compileApply funClosure argClosure generates code which causes the program to enter the closure represented by funClosure with the argument given by argClosure.

lookupVar :: (MonadCodeGen m, MonadIRBuilder m, MonadFail m) => Text -> Type -> m Operand

lookupVar name type generates code which retrieves the variable named name from the local environment. If the variable is not in scope, we generate code which prints an error at runtime. For convenience, we cast the type of the resulting Operand to type, which will normally be closureTyPtr.

Misc

retClosure :: (MonadModuleBuilder m, MonadIRBuilder m) => ClosurePtr k -> m ()

retClosure closurePtr returns the pointer represented by closurePtr.