Hachi.Compiler.CodeGen.Common

Description

This module contains common code generation functions.

Documentation

mkParamName :: Text -> ParameterName

mkParamName name converts a Text value into a ParameterName.

fatal :: (MonadModuleBuilder m, MonadIRBuilder m) => Operand -> [Operand] -> m ()

fatal strRef args generates code which terminates execution of the resulting program after displaying the string pointed to by strRef. Optionally, if the string contains formatting characters, args are used as values for those placeholders.

ifTracing :: MonadCodeGen m => m () -> m ()

ifTracing action will execute action if tracing is enabled.

compileTrace :: (MonadCodeGen m, MonadIRBuilder m) => String -> [Operand] -> m ()

compileTrace message arguments generates code which prints message to the standard output. The message string may contain formatting characters which will be substituted with values from arguments. If tracing is not enabled in the code generation configuration, no code is generated.