Hachi.Compiler.CodeGen.Constant

Description

This module contains code generation functions for constants.

Documentation

generateConstantGlobals :: (MonadCodeGen m, MonadFail m) => m ()

generateConstantGlobals is a computation which emits global definitions related to constants.

forceErrRef :: Operand

class CompileConstant a

Minimal complete definition

constantTypeId, compileConstant, compileLoadConstant, compilePrintBody

compileConst :: (MonadCodeGen m, MonadIRBuilder m) => Some (ValueOf DefaultUni) -> m (ClosurePtr 'StaticPtr)

compileConst value generates code for a constant whose value is provided as the first argument.

compileConstDynamic :: forall a m. (MonadCodeGen m, MonadIRBuilder m, CompileConstant a) => Operand -> m (ClosurePtr 'DynamicPtr)

compileConstDynamic value generates code which allocates a dynamic closure for the value given by value. Note that this function must be used with XTypeApplications in order to select a type for the type variable a.

retConstDynamic :: forall a m. (MonadCodeGen m, MonadIRBuilder m, CompileConstant a) => Continuation -> Operand -> m ()

retConstDynamic value is like compileConstDynamic, except that a return instruction is inserted at the end which returns the pointer to the new closure.

loadConstVal :: (MonadModuleBuilder m, MonadIRBuilder m) => Type -> m Operand

loadConstVal type loads the value of the result register (which constant closures write to when entered) assuming that it is of the type given by type.