Hachi.Compiler.CodeGen.Constant.ByteString

Documentation

bsNewStruct :: (MonadCodeGen m, MonadIRBuilder m) => Operand -> m Operand

bsNewStruct size generates code which allocates enough space for a new bytestring structure and stores size in it. The pointer to the new structure is returned. The data pointer of the bytestring is not set and no memory is allocated for the data.

bsNew :: (MonadCodeGen m, MonadIRBuilder m) => Operand -> m Operand

bsNew size generates code which allocates enough space for a new bytestring with size-many elements. The actual memory allocated is greater than size, since we need to store the size in the bytestring as well. This function performs the latter task as well before returning the pointer to the new bytestring.

bsLen :: (MonadCodeGen m, MonadIRBuilder m) => Operand -> m Operand

bsLen ptr generates code which loads the length of a bytestring which is represented by ptr.

bsDataPtr :: (MonadCodeGen m, MonadIRBuilder m) => Operand -> m Operand

bsDataPtr ptr generates code which calculates the address of the data component of a bytestring represented by ptr.

bsPrint :: (MonadCodeGen m, MonadIRBuilder m) => Operand -> m ()

bsPrint str generates code which prints the bytestring pointed at by str as a hexadecimal string to the standard output.

bsIndex :: (MonadCodeGen m, MonadIRBuilder m) => Operand -> Operand -> m Operand

bsIndex str ix checks that ix is a valid index for the bytestring pointed at by str and, if so, retrieves the character at the given index.

bsEquals :: (MonadCodeGen m, MonadIRBuilder m) => Operand -> Operand -> m Operand

bsEquals s0 s1 determines whether the two bytestrings pointed at by s0 and s1 are equal or not.

bsLessThan :: (MonadCodeGen m, MonadIRBuilder m) => Operand -> Operand -> m Operand

bsLessThanEquals :: (MonadCodeGen m, MonadIRBuilder m) => Operand -> Operand -> m Operand