Hachi.Compiler.CodeGen.Constant.Data

Tags

data DataTag

Constructors

DataConstr

DataMap

DataList

DataI

DataB

Code generation

dataGlobal :: MonadModuleBuilder m => Name -> DataTag -> [Constant] -> m Constant

dataGlobal name tag params constructs a new Data global named name, for the constructor identified by tag and with the parameters given by params.

newData :: (MonadCodeGen m, MonadIRBuilder m) => DataTag -> Operand -> Maybe Operand -> m Operand

newData tag dataPtr extraData allocates a new Data value on the heap for the constructor identified by tag where the data pointer is given by dataPtr and an additional argument by extraData which is normally empty or an extra tag in the case of DataConstr.

withDataTag :: (MonadModuleBuilder m, MonadIRBuilder m) => Operand -> (DataTag -> m ()) -> m ()

withDataTag ptr builder generates code which loads the tag from ptr and then performs different actions, produced by builder, depending on the tag value.

loadDataTag :: (MonadModuleBuilder m, MonadIRBuilder m) => Operand -> m Operand

loadDataTag ptr retrieves the data tag from ptr.

loadDataPtr :: (MonadModuleBuilder m, MonadIRBuilder m) => Operand -> m (ClosurePtr 'DynamicPtr)

loadDataPtr ptr retrieves the payload from the Data value pointed at by ptr. This function works for all forms of Data value.

loadConstrTag :: (MonadModuleBuilder m, MonadIRBuilder m) => Operand -> m (ClosurePtr 'DynamicPtr)

loadConstrTag ptr retrieves the constructor tag from a Data value pointed at by ptr. This function assumes that the Data value represents a data constructor and no such check is performed.