Hachi.Compiler.CodeGen.Constant.List

Documentation

listNew :: (MonadCodeGen m, MonadIRBuilder m) => Operand -> Operand -> m Operand

listNew head tail generates code which allocates enough space for a new list structure. head and tail are then stored in this structure.

getHead :: (MonadModuleBuilder m, MonadIRBuilder m) => Operand -> m (ClosurePtr 'DynamicPtr)

getHead listPtr generates code which retrieves the head of the list pointed at by listPtr.

getTail :: (MonadModuleBuilder m, MonadIRBuilder m) => Operand -> m (ClosurePtr 'DynamicPtr)

getTail listPtr generates code which retrieves the tail of the list pointed at by listPtr.

listCase :: MonadIRBuilder m => Operand -> m () -> m () -> m ()

listCase listPtr nullBuilder consBuilder generates code which tests if listPtr is a pointer to null or to a cons cell and uses nullBuilder and consBuilder to generate code for the respective cases.