Hachi.Compiler.CodeGen.Constant.Pair

Documentation

newPair :: (MonadCodeGen m, MonadIRBuilder m) => ClosurePtr p -> ClosurePtr q -> m Operand

newPair fst snd generates code which constructs a new pair value, where the first component is fst and the second component is snd.

getFst :: (MonadModuleBuilder m, MonadIRBuilder m) => Operand -> m (ClosurePtr 'DynamicPtr)

getFst pairPtr generates code which retrieves the first component from the pair pointed at by pairPtr.

getSnd :: (MonadModuleBuilder m, MonadIRBuilder m) => Operand -> m (ClosurePtr 'DynamicPtr)

getSnd pairPtr generates code which retrieves the second component from the pair pointed at by pairPtr.