Hachi.Compiler.CodeGen.Driver

Documentation

runDriver :: Config -> [String] -> [String] -> IO ()

runDriver config pkgConfigArgs clangArgs runs Clang with arguments given by clangArgs and additional arguments obtained by running pkg-config with the arguments given by pkgConfigArgs for all of our dependencies.

buildObjectFile :: Config -> FilePath -> IO FilePath

buildObjectFile config sourceFile uses Clang to compile the source file at sourceFile to an object file. The path of the generated object file is returned.

linkExecutable :: Config -> FilePath -> [FilePath] -> IO ()

linkExecutable config outputName objectFiles runs Clang to produce an executable using the configuration given by config, where outputName is the path we computed for the output, and objectFiles are the paths of the object files that we want to link together.