Hachi.Compiler.CodeGen.Externals

Description

This module contains the types, declarations, and references to external functions that we import from stdlib.

Documentation

externalDefinitions :: [Definition]

rtsInit :: (MonadModuleBuilder m, MonadIRBuilder m) => m ()

Standard C library

printfTy :: Type

The type of printf.

printfRef :: Constant

A Constant which provides a reference to the global printf declaration.

printf :: (MonadModuleBuilder m, MonadIRBuilder m) => Operand -> [Operand] -> m Operand

exitTy :: Type

exitRef :: Constant

exit :: (MonadModuleBuilder m, MonadIRBuilder m) => Integer -> m Operand

mallocTy :: Type

mallocRef :: Constant

malloc :: (MonadModuleBuilder m, MonadIRBuilder m, MonadCodeGen m) => Type -> Operand -> m Operand

malloc type size generates code which calls malloc to allocate size-many bytes. The return value is cast to type, which should normally be a pointer to something.

memcpyTy :: Type

memcpyRef :: Constant

memcpy :: (MonadModuleBuilder m, MonadIRBuilder m) => Operand -> Operand -> Operand -> m Operand

memcmp :: (MonadModuleBuilder m, MonadIRBuilder m) => Operand -> Operand -> Operand -> m Operand

Cryptography

sha2_256 :: (MonadModuleBuilder m, MonadIRBuilder m) => Operand -> m Operand

sha3_256 :: (MonadModuleBuilder m, MonadIRBuilder m) => Operand -> m Operand

blake2b :: (MonadModuleBuilder m, MonadIRBuilder m) => Operand -> m Operand

verifySig :: (MonadModuleBuilder m, MonadIRBuilder m) => Operand -> Operand -> Operand -> m Operand

GMP

mpzInitSetStr :: (MonadModuleBuilder m, MonadIRBuilder m) => Name -> Constant -> String -> m ()

mpzInitSetUInt :: (MonadModuleBuilder m, MonadIRBuilder m) => Operand -> Operand -> m ()

mpzGetStr :: (MonadModuleBuilder m, MonadIRBuilder m) => Operand -> Operand -> Operand -> m Operand

mpzGetUInt :: (MonadModuleBuilder m, MonadIRBuilder m) => Operand -> m Operand

mpzAdd :: (MonadModuleBuilder m, MonadIRBuilder m) => Operand -> Operand -> Operand -> m ()

mpzSub :: (MonadModuleBuilder m, MonadIRBuilder m) => Operand -> Operand -> Operand -> m ()

mpzMul :: (MonadModuleBuilder m, MonadIRBuilder m) => Operand -> Operand -> Operand -> m ()

mpzFDivQ :: (MonadModuleBuilder m, MonadIRBuilder m) => Operand -> Operand -> Operand -> m ()

mpzFDivR :: (MonadModuleBuilder m, MonadIRBuilder m) => Operand -> Operand -> Operand -> m ()

mpzTDivQ :: (MonadModuleBuilder m, MonadIRBuilder m) => Operand -> Operand -> Operand -> m ()

mpzTDivR :: (MonadModuleBuilder m, MonadIRBuilder m) => Operand -> Operand -> Operand -> m ()

mpzCmp :: (MonadModuleBuilder m, MonadIRBuilder m) => Operand -> Operand -> m Operand