Hachi.Compiler.CodeGen.IRBuilder

Description

This module provides convenience functions for constructing LLVM IR, including more flexible versions of functions found in IRBuilder.

Re-exports

array :: (HasCallStack, MonadModuleBuilder m) => [Constant] -> m Operand

struct :: Maybe Name -> Bool -> [Constant] -> Operand

half :: Word16 -> Operand

single :: Float -> Operand

double :: Double -> Operand

bit :: Integer -> Operand

int8 :: Integer -> Operand

int16 :: Integer -> Operand

int32 :: Integer -> Operand

int64 :: Integer -> Operand

sizeof :: (HasCallStack, MonadIRBuilder m, MonadModuleBuilder m) => Word32 -> Type -> m Operand

unreachable :: MonadIRBuilder m => m ()

See reference.

condBr :: MonadIRBuilder m => Operand -> Name -> Name -> m ()

Conditional branch (see br for unconditional instructions). See reference.

select :: (HasCallStack, MonadIRBuilder m, MonadModuleBuilder m) => Operand -> Operand -> Operand -> m Operand

See reference.

switch :: MonadIRBuilder m => Operand -> Name -> [(Constant, Name)] -> m ()

See reference.

ret :: MonadIRBuilder m => Operand -> m ()

See reference.

retVoid :: MonadIRBuilder m => m ()

Emit a ret void instruction. See reference.

phi :: (HasCallStack, MonadIRBuilder m, MonadModuleBuilder m) => [(Operand, Name)] -> m Operand

See reference.

br :: MonadIRBuilder m => Name -> m ()

Unconditional branch. Emit a br label dest instruction See reference.

fcmp :: MonadIRBuilder m => FloatingPointPredicate -> Operand -> Operand -> m Operand

See reference.

icmp :: MonadIRBuilder m => IntegerPredicate -> Operand -> Operand -> m Operand

See reference.

insertValue :: (HasCallStack, MonadIRBuilder m, MonadModuleBuilder m) => Operand -> Operand -> [Word32] -> m Operand

See reference.

extractValue :: (HasCallStack, MonadIRBuilder m, MonadModuleBuilder m) => Operand -> [Word32] -> m Operand

See reference.

shuffleVector :: (HasCallStack, MonadIRBuilder m, MonadModuleBuilder m) => Operand -> Operand -> [Int32] -> m Operand

See reference.

insertElement :: (HasCallStack, MonadIRBuilder m, MonadModuleBuilder m) => Operand -> Operand -> Operand -> m Operand

See reference.

extractElement :: (HasCallStack, MonadIRBuilder m, MonadModuleBuilder m) => Operand -> Operand -> m Operand

See reference.

bitcast :: MonadIRBuilder m => Operand -> Type -> m Operand

Emit the bitcast ... to instruction. See reference.

inttoptr :: MonadIRBuilder m => Operand -> Type -> m Operand

Emit the inttoptr ... to instruction. See reference.

ptrtoint :: MonadIRBuilder m => Operand -> Type -> m Operand

Emit the ptrtoint ... to instruction. See reference.

sitofp :: MonadIRBuilder m => Operand -> Type -> m Operand

Emit the sitofp ... to instruction. See reference.

uitofp :: MonadIRBuilder m => Operand -> Type -> m Operand

Emit the uitofp ... to instruction. See reference.

fpext :: MonadIRBuilder m => Operand -> Type -> m Operand

Emit the fpext ... to instruction. See reference.

fptosi :: MonadIRBuilder m => Operand -> Type -> m Operand

Emit the fptosi ... to instruction. See reference.

fptoui :: MonadIRBuilder m => Operand -> Type -> m Operand

Emit the fptoui ... to instruction. See reference.

sext :: MonadIRBuilder m => Operand -> Type -> m Operand

Emit the sext ... to instruction. See reference.

zext :: MonadIRBuilder m => Operand -> Type -> m Operand

Emit the zext ... to instruction. See reference.

fptrunc :: MonadIRBuilder m => Operand -> Type -> m Operand

Emit the fptrunc ... to instruction. See reference.

trunc :: MonadIRBuilder m => Operand -> Type -> m Operand

Emit the trunc ... to instruction. See reference.

gep :: (HasCallStack, MonadIRBuilder m, MonadModuleBuilder m) => Operand -> [Operand] -> m Operand

Emit the getelementptr instruction. See reference.

store :: MonadIRBuilder m => Operand -> Word32 -> Operand -> m ()

See reference.

load :: (HasCallStack, MonadIRBuilder m, MonadModuleBuilder m) => Operand -> Word32 -> m Operand

See reference.

alloca :: MonadIRBuilder m => Type -> Maybe Operand -> Word32 -> m Operand

See reference.

xor :: (MonadIRBuilder m, MonadModuleBuilder m) => Operand -> Operand -> m Operand

See reference.

or :: (MonadIRBuilder m, MonadModuleBuilder m) => Operand -> Operand -> m Operand

See reference.

and :: (MonadIRBuilder m, MonadModuleBuilder m) => Operand -> Operand -> m Operand

See reference.

ashr :: (MonadIRBuilder m, MonadModuleBuilder m) => Operand -> Operand -> m Operand

See reference.

lshr :: (MonadIRBuilder m, MonadModuleBuilder m) => Operand -> Operand -> m Operand

See reference.

shl :: (MonadIRBuilder m, MonadModuleBuilder m) => Operand -> Operand -> m Operand

See reference.

srem :: (MonadIRBuilder m, MonadModuleBuilder m) => Operand -> Operand -> m Operand

See reference.

urem :: (MonadIRBuilder m, MonadModuleBuilder m) => Operand -> Operand -> m Operand

See reference.

sdiv :: (MonadIRBuilder m, MonadModuleBuilder m) => Operand -> Operand -> m Operand

See reference.

udiv :: (MonadIRBuilder m, MonadModuleBuilder m) => Operand -> Operand -> m Operand

See reference.

sub :: (MonadIRBuilder m, MonadModuleBuilder m) => Operand -> Operand -> m Operand

See reference.

mul :: (MonadIRBuilder m, MonadModuleBuilder m) => Operand -> Operand -> m Operand

See reference.

add :: (MonadIRBuilder m, MonadModuleBuilder m) => Operand -> Operand -> m Operand

See reference.

frem :: (MonadIRBuilder m, MonadModuleBuilder m) => Operand -> Operand -> m Operand

See reference.

fdiv :: (MonadIRBuilder m, MonadModuleBuilder m) => Operand -> Operand -> m Operand

See reference.

fsub :: (MonadIRBuilder m, MonadModuleBuilder m) => Operand -> Operand -> m Operand

See reference.

fmul :: (MonadIRBuilder m, MonadModuleBuilder m) => Operand -> Operand -> m Operand

See reference.

fadd :: (MonadIRBuilder m, MonadModuleBuilder m) => Operand -> Operand -> m Operand

See reference.

fneg :: (MonadIRBuilder m, MonadModuleBuilder m) => Operand -> m Operand

See reference.

buildModuleT :: Monad m => ShortByteString -> ModuleBuilderT m a -> m Module

Convenience function for module construction (transformer version)

buildModule :: ShortByteString -> ModuleBuilder a -> Module

Convenience function for module construction

typedef :: MonadModuleBuilder m => Name -> Maybe Type -> m Type

A named type definition

externVarArgs

Arguments

:: MonadModuleBuilder m

=> Name
Definition name

-> [Type]
Parameter types

-> Type
Type

-> m Operand

An external variadic argument function definition

extern

Arguments

:: MonadModuleBuilder m

=> Name
Definition name

-> [Type]
Parameter types

-> Type
Type

-> m Operand

An external function definition

emitDefn :: MonadModuleBuilder m => Definition -> m ()

execModuleBuilderT :: Monad m => ModuleBuilderState -> ModuleBuilderT m a -> m [Definition]

Evaluate ModuleBuilderT to a list of definitions

execModuleBuilder :: ModuleBuilderState -> ModuleBuilder a -> [Definition]

Evaluate ModuleBuilder to a list of definitions

runModuleBuilderT :: Monad m => ModuleBuilderState -> ModuleBuilderT m a -> m (a, [Definition])

Evaluate ModuleBuilderT to a result and a list of definitions

runModuleBuilder :: ModuleBuilderState -> ModuleBuilder a -> (a, [Definition])

Evaluate ModuleBuilder to a result and a list of definitions

emptyModuleBuilder :: ModuleBuilderState

newtype ModuleBuilderT (m :: Type -> Type) a

Constructors

ModuleBuilderT

Fields

unModuleBuilderT :: StateT ModuleBuilderState m a

data ModuleBuilderState

Constructors

ModuleBuilderState

Fields

builderDefs :: SnocList Definition

builderTypeDefs :: Map Name Type

type ModuleBuilder = ModuleBuilderT Identity

class Monad m => MonadModuleBuilder (m :: Type -> Type) where

Minimal complete definition

Nothing

Methods

liftModuleState :: State ModuleBuilderState a -> m a

data ParameterName

A parameter name suggestion

Constructors

NoParameterName

ParameterName ShortByteString

hasTerminator :: (HasCallStack, MonadIRBuilder m) => m Bool

Find out if the currently active block has a terminator.

This function will fail under the same condition as currentBlock

currentBlock :: (HasCallStack, MonadIRBuilder m) => m Name

Get the name of the currently active block.

This function will throw an error if there is no active block. The only situation in which this can occur is if it is called before any call to block and before emitting any instructions.

named :: MonadIRBuilder m => m r -> ShortByteString -> m r

ir named name executes the IRBuilder ir using name as the base name whenever a fresh local name is generated. Collisions are avoided by appending numbers (first "name", then "name1", "name2", and so on).

block :: MonadIRBuilder m => m Name

Starts a new block and ends the previous one

emitBlockStart :: MonadIRBuilder m => Name -> m ()

Starts a new block labelled using the given name and ends the previous one. The name is assumed to be fresh.

emitTerm :: MonadIRBuilder m => Terminator -> m ()

Emit terminator

emitInstrVoid :: MonadIRBuilder m => Instruction -> m ()

Emit instruction that returns void

emitInstr

Arguments

:: MonadIRBuilder m

=> Type
Return type

-> Instruction

-> m Operand

Emit instruction

freshUnName :: MonadIRBuilder m => m Name

Generate a fresh numbered name

freshName :: MonadIRBuilder m => ShortByteString -> m Name

Generate a fresh name from a name suggestion

fresh :: MonadIRBuilder m => m Name

Generate a fresh name. The resulting name is numbered or based on the name suggested with named if that's used.

modifyBlock :: MonadIRBuilder m => (PartialBlock -> PartialBlock) -> m ()

ensureBlock :: MonadIRBuilder m => m ()

If no partial block exists, create a new block with a fresh label.

This is useful if you want to ensure that the label for the block is assigned before another label which is not possible with modifyBlock.

execIRBuilderT :: Monad m => IRBuilderState -> IRBuilderT m a -> m [BasicBlock]

Evaluate IRBuilderT to a list of basic blocks

execIRBuilder :: IRBuilderState -> IRBuilder a -> [BasicBlock]

Evaluate IRBuilder to a list of basic blocks

runIRBuilderT :: Monad m => IRBuilderState -> IRBuilderT m a -> m (a, [BasicBlock])

Evaluate IRBuilderT to a result and a list of basic blocks

runIRBuilder :: IRBuilderState -> IRBuilder a -> (a, [BasicBlock])

Evaluate IRBuilder to a result and a list of basic blocks

emptyIRBuilder :: IRBuilderState

emptyPartialBlock :: Name -> PartialBlock

newtype IRBuilderT (m :: Type -> Type) a

This provides a uniform API for creating instructions and inserting them into a basic block: either at the end of a BasicBlock, or at a specific location in a block.

Constructors

IRBuilderT

Fields

unIRBuilderT :: StateT IRBuilderState m a

type IRBuilder = IRBuilderT Identity

class Monad m => MonadIRBuilder (m :: Type -> Type) where

Minimal complete definition

Nothing

Methods

liftIRState :: State IRBuilderState a -> m a

data PartialBlock

A partially constructed block as a sequence of instructions

Constructors

PartialBlock

Fields

partialBlockName :: !Name

partialBlockInstrs :: SnocList (Named Instruction)

partialBlockTerm :: First (Named Terminator)

data IRBuilderState

Builder monad state

Constructors

IRBuilderState

Fields

builderSupply :: !Word

builderUsedNames :: !(Map ShortByteString Word)

builderNameSuggestion :: !(Maybe ShortByteString)

builderBlocks :: SnocList BasicBlock

builderBlock :: !(Maybe PartialBlock)

Modifiers

setLinkage :: Linkage -> Global -> Global

setLinkage linkage global sets global's linkage type to linkage.

class HasCC a where

Methods

setCC :: CallingConvention -> a -> a

setCC cc object sets the calling convention of object to cc.

class HasReturnAttr a where

Methods

setReturnAttr :: [ParameterAttribute] -> a -> a

setReturnAttr attributes object sets the return attributes of object to attributes.

class HasFunctionAttr a where

Methods

setFunctionAttr :: [Either GroupID FunctionAttribute] -> a -> a

setFunctionAttr attributes object sets the function attributes of object to attributes.

class HasMetadata a where

Methods

setMetadata :: [(ShortByteString, MDRef MDNode)] -> a -> a

setMetadata metadata object sets the metadata of object to metadata.

plcFunOpts :: Global -> Global

plcFunOpts global applies the default function settings we use for PLC code to global.

Globals

global :: MonadModuleBuilder m => Name -> Type -> Constant -> (Global -> Global) -> m Operand

global name type initialValue transformation emits a new global named name of type with an initial value of initialValue. Additionally, transformation is applied to the global before it is emitted, so that arbitrary transformations can be performed on it.

globalStringPtr :: (MonadModuleBuilder m, MonadFail m) => String -> Name -> (Global -> Global) -> m Constant

globalStringPtr string name transformation emits a new global string constant named name with a value of string. The transformation is applied to the Global before it is emitted.

function :: MonadModuleBuilder m => Name -> [(Type, ParameterName)] -> Type -> (Global -> Global) -> ([Operand] -> IRBuilderT m ()) -> m Operand

function name parameters returnType transformation bodyBuilder emits a new function named name.

Instructions

plcCall :: Instruction -> Instruction

plcCall instruction modified a Call instruction so that it is suitable for calls to PLC functions.

call :: (HasCallStack, MonadFail m, MonadIRBuilder m, MonadModuleBuilder m) => Operand -> [(Operand, [ParameterAttribute])] -> (Instruction -> Instruction) -> m Operand

call function arguments transformation emits a new Call instruction for a call to function with arguments. The transformation is applied to the Call instruction before it is emitted.

castIfNeeded :: (MonadModuleBuilder m, MonadIRBuilder m) => Type -> Operand -> m Operand

castIfNeeded type ptr casts the type of ptr to type, but only if it is not already type.

castToClosure :: (MonadModuleBuilder m, MonadIRBuilder m) => Operand -> m Operand

castToClosure ptr casts the type of ptr to closureTyPtr, but only if it is not already closureTyPtr.