Hachi.Compiler.CodeGen.Library

Description

Code generation for functions we want to export when compiling to a library.

Documentation

emitLibraryApi :: MonadCodeGen m => m String

emitLibraryApi is a computation which generates all functions that should be available as part of the library API. It also returns a String value containing all of the C function signatures for the API.