Hachi.Compiler.CodeGen.Monad

Documentation

data ConstantTy

Enumerates all supported types of constants. This would technically be a great candidate for a data family, but we'd then end up with some mutually recursive modules, unless we shuffle things around a bunch.

Constructors

ConstInteger

ConstByteString

ConstText

ConstUnit

ConstBool

ConstData

ConstPair

ConstList

data CodeGenSt

Represents the state of the code generator.

Constructors

MkCodeGenSt

Fields

codeGenCfg :: Config

The configuration for the compiler.

codeGenErrMsg :: Constant

codeGenCounters :: IORef (Map String (IORef Integer))

codeGenEnv :: Map Text Local

codeGenBuiltins :: Map DefaultFun (ClosurePtr 'StaticPtr)

codeGenConstEntries :: IORef (Map ConstantTy Constant)

codeGenConstPrinters :: IORef (Map ConstantTy Constant)

codeGenFunPrinter :: IORef (Maybe Constant)

codeGenFreeVars :: Set (Text, Bool)

newtype CodeGen m a

The code generator monad.

Constructors

MkCodeGen

Fields

runCodeGen :: ReaderT CodeGenSt m a

type MonadCodeGen m = (MonadIO m, MonadFail m, MonadModuleBuilder m, MonadReader CodeGenSt m, MonadFix m)

A constraint synonym for the type class constraints we usually expect our code generation monad to satisfy.

mkFresh :: (MonadReader CodeGenSt m, MonadIO m) => String -> m String

mkFresh prefix generates a fresh name starting with prefix.

extendScope :: MonadReader CodeGenSt m => Text -> Local -> m a -> m a

extendScope name operand action adds name to the scope for action along with operand which represents the LLVM identifier.

updateEnv :: MonadReader CodeGenSt m => Set Text -> [Local] -> m a -> m a

updateEnv names operands action updates the scope for action with operands which correspond to names. Important: the names in names must be in the same order as the corresponding entries in operands.

localFreeVars :: MonadReader CodeGenSt m => Set (Text, Bool) -> m a -> m a

localFreeVars freeVars sets the set of free variables in the current compilation scope to freeVars.

currentFreeVars :: MonadReader CodeGenSt m => m (Set (Text, Bool))

currentFreeVars is a computation which retrieves the set of free variables in the current compilation scope.