Hachi.Compiler.CodeGen.Types

Description

This module contains various LLVM Type values used in the code generator.

Re-exports

module Hachi.Compiler.Platform

Utility functions

isPtr :: Type -> Bool

isPtr type determines whether type represents a pointer type.

Basic LLVM types

i1 :: Type

i8 :: Type

i32 :: Type

i64 :: Type

ptrOf :: Type -> Type

char :: Type

stringPtr :: String -> Type

asStringPtr :: Constant -> Constant

asStringPtr constant casts constant to a char pointer.

Bytestrings

bytestringTyDef :: Type

bytestringTyDef is the type definition for bytestrings.

bytestringTy :: Type

bytestringTy is a Type for bytestrings.

bytestringTyPtr :: Type

bytestringTyPtr is a Type representing a pointer to a bytestring.

Pairs

pairTyDef :: Type

pairTyDef is the type definition for pairs.

pairTy :: Type

pairTy is a Type for pairs.

pairTyPtr :: Type

pairTyPtr is a Type representing a pointer to a pair.

Lists

listTyDef :: Type

listTyDef is the type definition for lists.

listTy :: Type

listTy is a Type for lists.

listTyPtr :: Type

listTyPtr is a Type representing a pointer to a list.

newtype ListPtr

Represents a pointer to a linked list.

Constructors

MkListPtr

Fields

listPtr :: Operand

Data

dataTyDef :: Type

The Plutus Data type is a sum type with five different constructors. Our representation is as a tagged array of pointers: For Constr, we have a constructor tag and a pointer to a list structure For Map, we have a pointer to a list structure For List, we have a pointer to a list structure For I, we have an integer value For B, we have a pointer to a bytestring structure

dataTy :: Type

dataTy is a Type for data values.

dataTyPtr :: Type

dataTyPtr is a Type representing a pointer to a data value.

GMP

gmpTyDef :: Type

gmpTyDef attempts to capture the structure of __mpz_struct from the gmp library.

emptyGmpTy :: Constant

gmpTy :: Type

gmpTy is equivalent to mpz_t from the gmp library.

gmpTyPtr :: Type

gmpTyPtr is a pointer to gmpTy.

Closures

closureTy :: Type

closureTy is a Type for closures.

closureTyPtr :: Type

closureTyPtr is a Type representing a pointer to a closure.

data PtrKind

Enumerates different kinds of pointers we may have:

StaticPtr is guaranteed to represent a pointer to a static global

DynamicPtr _may_ represent a pointer to a dynamically allocated value.

In other words, we use StaticPtr when we know what sort of pointer we have and DynamicPtr in other cases.

Constructors

StaticPtr

DynamicPtr

data ClosurePtr (k :: PtrKind) where

Represents a pointer to a closure.

Constructors

MkClosurePtr :: Operand -> ClosurePtr 'DynamicPtr

MkStaticClosurePtr :: Constant -> ClosurePtr 'StaticPtr

class FromClosurePtr k to where

Methods

closurePtr :: ClosurePtr k -> to

closurePtr closurePtr returns closurePtr's underlying LLVM representation.

toDynamicPtr :: ClosurePtr k -> ClosurePtr 'DynamicPtr

toDynamicPtr closurePtr turns a closure pointer into a dynamic closure pointer. This is conceptually a no-op and just allows us to forget information in case we only need a DynamicPtr, but have a StaticPtr or don't know which one we have. Static pointers are always acceptable where dynamic pointers are expected.

Continuations

contTy :: Type

contTy is a Type for continuations.

contTyPtr :: Type

contTyPtr is a Type representing a pointer to a continuation.

newtype Continuation

Represents pointers to continuations.

Constructors

MkCont

Fields

contPtr :: Operand

Locals

data Local

Local variables can be closures or continuations.

Constructors

LocalClosure (ClosurePtr 'DynamicPtr)

LocalCont Continuation

localOperand :: Local -> Operand

localOperand local retrieves the underlying Operand of local.