Hachi.Compiler.Config

Documentation

data Config

Constructors

MkConfig

Fields

cfgInput :: FilePath

cfgOutput :: Maybe FilePath

cfgRTS :: Maybe FilePath

cfgDeserialise :: Bool

Is the input a binary file that must be deserialised?

cfgTyped :: Bool

Is the input typed?

cfgTrace :: Bool

Should the resulting LLVM code print diagnostic information?

cfgVerbose :: Bool

Should the compiler print diagnostic information while running?

cfgNoAssemble :: Bool

Should the compiler generate an object file?

cfgNoLink :: Bool

Should the compiler generate an executable?

cfgNoSerialise :: Bool

Should the LLVM IR be in plain-text?

cfgCheckOOM :: Bool

Should the generated code check for OOM after every malloc?

cfgLibrary :: Bool

Should we produce a library?

cfgEntryPoint :: Maybe FilePath

If --library is specified, an optional path to a .c source file which contains the entry point for the program.

cfgOptimisationLevel :: Maybe Word

An optional value indicating what set of optimisations should be applied to the program.

mkDefaultConfig :: FilePath -> Config

mkDefaultConfig input constructs a Config with reasonable defaults and using input as the input source file.

parseCmdLineArgs :: IO Config

parseCmdLineArgs parses command-line arguments into a Config value. If parsing fails, the program is terminated and a help message is shown.