Hachi.Compiler.FreeVars

Documentation

freeVars :: Ord name => Term name uni fun ann -> Set name

freeVars term calculates the set of free variables for term. For some reason, this function does not seem to exist in the plutus libraries (it does exist for typed plutus and plutus IR though).